+++ test/doing_output_test.rb
@@ -10,6 +10,8 @@ require 'test_helper'
 class DoingOutputTest < Test::Unit::TestCase
   include DoingHelpers
   ENTRY_REGEX = /^\d{4}-\d\d-\d\d \d\d:\d\d \|/.freeze
+  ENTRY_TS_REGEX = /\s*(?<ts>[^|]+) \s*\|/.freeze
+  ENTRY_DONE_REGEX = /@done\((?<ts>.*?)\)/.freeze
 
   def setup
     @tmpdirs = []
@@ -17,81 +19,12 @@ class DoingOutputTest < Test::Unit::TestCase
     @basedir = mktmpdir
     @wwid_file = File.join(@basedir, 'wwid.md')
     @config_file = File.join(File.dirname(__FILE__), 'test.doingrc')
+    @import_file = File.join(File.dirname(__FILE__), 'All Activities 2.json')
     @config = YAML.load(IO.read(@config_file))
   end
 
   def teardown
     FileUtils.rm_rf(@tmpdirs)
-  end
-
-  def test_last_command
-    subject = 'Test new entry @tag1'
-    doing('now', subject)
-    assert_match(/#{subject}\s*$/, doing('last'), 'last entry should be entry just added')
-  end
-
-  def test_last_search_and_tag
-    unique_keyword = 'jumping jesus'
-    unique_tag = 'balloonpants'
-    doing('now', "Test new entry @#{unique_tag} sad monkey")
-    doing('now', "Test new entry @tag2 #{unique_keyword}")
-    doing('now', 'Test new entry @tag3 burly man')
-
-    assert_match(/#{unique_keyword}/, doing('last', '--search', unique_keyword), 'returned entry should contain unique keyword')
-    assert_match(/@#{unique_tag}/, doing('last', '--tag', unique_tag), 'returned entry should contain unique tag')
-  end
-
-  def test_view_command
-    subject = 'Test new entry @tag1'
-    doing('done', subject)
-    subject2 = 'Test new entry 2 @tag2'
-    doing('done', subject2)
-    assert_count_entries(2, doing('view', 'done'), 'There should be 2 entries shown by `view done`')
-  end
-
-  def test_show_command
-    subject = 'Test new entry @tag1'
-    doing('now', subject)
-    subject2 = 'Test new entry 2 @tag2'
-    doing('now', subject2)
-    result = doing('show').uncolor.strip
-    assert_count_entries(2, result, 'There should be 2 entries shown by `doing show`')
-    assert_match(/#{subject}\s*$/, result, 'doing show results should include test entry')
-    result = doing('show', '@tag1').uncolor.strip
-    assert_count_entries(1, result, 'There should be 1 entries shown by `doing show @tag1`')
-    assert_match(/#{subject}\s*$/, result, 'doing show @tag1 results should include test entry')
-  end
-
-  def test_show_command_tag_boolean
-    subject = 'Test new entry @tag1'
-    doing('now', subject)
-    subject2 = 'Test new entry 2 @tag2'
-    doing('now', subject2)
-    subject3 = 'Test new entry 3 @tag1 @tag2 @tag3'
-    doing('now', subject3)
-
-    result = doing('show', '--tag', 'tag1,tag2', '--bool', 'and').uncolor.strip
-    assert_count_entries(1, result, 'There should be 1 entry shown with both @tag1 and @tag2')
-    assert_match(/#{subject3}\s*$/, result, 'doing show results should include entry with both @tag1 and @tag2')
-
-    result = doing('show', '--tag', 'tag1,tag2', '--bool', 'or').uncolor.strip
-    assert_count_entries(3, result, 'There should be 3 entries shown with either @tag1 or @tag2')
-    result = doing('show', '--tag', 'tag2', '--bool', 'not').uncolor.strip
-    assert_count_entries(1, result, 'There should be 1 entry shown without @tag2')
-    assert_match(/#{subject}\s*$/, result, 'doing show results should include entry without @tag2')
-  end
-
-  def test_today_command
-    subject = 'Test new entry @tag1'
-    doing('done', subject)
-    subject2 = 'Test new entry 2 @tag2'
-    doing('now', subject2)
-    assert_count_entries(2, doing('today'), 'There should be 2 entries shown by `doing today`')
-  end
-
-  def test_yesterday_command
-    doing('done', 'Adding an entry finished yesterday', '--took', '30m', '--back', 'yesterday 3pm')
-    assert_count_entries(1, doing('yesterday'), 'There should be 1 entry shown by `doing yesterday`')
   end
 
   def test_sections_command
@@ -109,14 +42,6 @@ class DoingOutputTest < Test::Unit::TestCase
     assert_equal(matches.count, 2, 'There should be 2 entries shown by `doing recent`')
   end
 
-  def test_on_command
-    # 1:42pm: Did a thing @done(2021-07-05 13:42)
-    doing('now', 'Test new entry @tag1')
-    doing('now', 'Test new entry 2 @tag2')
-    result = doing('--stdout', 'on', 'today')
-    assert_count_entries(2, result, 'There should be 2 entries')
-  end
-
   def test_template_command
     result = doing('template', 'haml')
     assert_match(/^!!!\s*\n%html/, result, 'Output should be a HAML template')
@@ -125,6 +50,16 @@ class DoingOutputTest < Test::Unit::TestCase
   end
 
   private
+
+  def assert_matches(matches, shown)
+    matches.each do |regexp, msg, opt_refute|
+      if opt_refute
+        assert_no_match(regexp, shown, msg)
+      else
+        assert_match(regexp, shown, msg)
+      end
+    end
+  end
 
   def assert_count_entries(count, shown, message = 'Should be X entries shown')
     assert_equal(count, shown.uncolor.strip.scan(ENTRY_REGEX).count, message)